Enemy 상태, Ramona 이미지 수정

+++ gameNode.h
@@ -21,7 +21,7 @@ public:
 	virtual void update();			//연산하는 함수
 	virtual void render();	//그리기 함수
 	
-	virtual int getRenderPosY(); // 렌더 포지션을 위한 함수
+	virtual float getRenderPosY(); // 렌더 포지션을 위한 함수
 
 	//백버퍼의 DC영역에 대한 접근자
 	HDC getMemDC() { return _backBuffer->getMemDC(); }
